requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.10.4
ortools==9.11.4210
rapidfuzz==3.10.1
Jinja2==3.1.4
streamlit==1.41.1
googlemaps==4.10.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
